PLCC-299 Move Term to root
