+++ crates/api/src/atom/atom_test.rs
@@ -5,7 +5,7 @@
 use approx::{assert_relative_eq, assert_ulps_eq};
 
 use crate::{
-    parser::term::UNITY_ARRAY, ucum_unit::UcumUnit, Atom, Classification, Composable, Composition,
+    term::UNITY_ARRAY, ucum_unit::UcumUnit, Atom, Classification, Composable, Composition,
     Dimension, UcumSymbol,
 };
