+++ crates/api/src/composition.rs
@@ -1,6 +1,6 @@
 use std::{fmt, ops::Mul};
 
-use crate::{parser::term::Exponent, Dimension};
+use crate::{term::Exponent, Dimension};
 
 pub const DIMLESS: Composition = Composition::new_dimless();
