+++ crates/api/src/lib.rs
@@ -37,6 +37,7 @@ pub mod parser;
 mod prefix;
 pub mod property;
 pub mod reduce;
+pub(crate) mod term;
 mod ucum_symbol;
 pub mod unit;
 
@@ -46,6 +47,6 @@ mod ucum_unit;
 pub use crate::{
     atom::Atom, classification::Classification, composable::Composable, composition::Composition,
     convertible::Convertible, dimension::Dimension, error::Error, field_eq::FieldEq,
-    is_compatible_with::IsCompatibleWith, measurement::Measurement, parser::Term, prefix::Prefix,
-    property::Property, ucum_symbol::UcumSymbol, ucum_unit::UcumUnit, unit::Unit,
+    is_compatible_with::IsCompatibleWith, measurement::Measurement, prefix::Prefix,
+    property::Property, term::Term, ucum_symbol::UcumSymbol, ucum_unit::UcumUnit, unit::Unit,
 };
