+++ crates/api/src/parser.rs
@@ -1,17 +1,16 @@
 #![allow(clippy::large_enum_variant)]
 #![allow(clippy::result_large_err)]
 
-mod annotation_composition;
+pub(crate) mod annotation_composition;
 mod error;
 mod symbols;
-pub(crate) mod term;
 mod terms;
 
 use pest::{iterators::Pair, Parser};
 
-use crate::{Atom, Prefix};
+use crate::{Atom, Prefix, Term};
 
-pub use self::{annotation_composition::AnnotationComposition, error::Error, term::Term};
+pub use self::{annotation_composition::AnnotationComposition, error::Error};
 
 use self::{
     symbols::symbol_parser::Rule as SymbolRule,
