+++ crates/api/src/parser/terms/mapper.rs
@@ -56,9 +56,11 @@ pub(crate) fn map(mut pairs: Pairs<'_, Rule>) -> Result<Vec<Term>, Error> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::parser::terms::term_parser::{Rule, TermParser};
     use pest::Parser;
+
+    use crate::parser::terms::term_parser::{Rule, TermParser};
+
+    use super::*;
 
     macro_rules! validate_interpret {
         ($test_name:ident, $input:expr, $($terms:expr),+) => {
