+++ crates/api/src/parser/terms/mapper/annotatable.rs
@@ -1,7 +1,8 @@
 use pest::iterators::Pair;
 
-use crate::parser::{
-    term::Exponent as IExponent, terms::term_parser::Rule as TermRule, Error, Visit,
+use crate::{
+    parser::{terms::term_parser::Rule as TermRule, Error, Visit},
+    term::Exponent as IExponent,
 };
 
 use super::{Atom, Exponent, Prefix, SimpleUnit};
