+++ crates/api/src/parser/terms/mapper/factor.rs
@@ -1,6 +1,9 @@
 use pest::iterators::Pair;
 
-use crate::parser::{term, terms::term_parser::Rule as TermRule, Error, Visit};
+use crate::{
+    parser::{terms::term_parser::Rule as TermRule, Error, Visit},
+    term,
+};
 
 impl Visit<TermRule> for term::Factor {
     fn visit(pair: Pair<'_, TermRule>) -> Result<Self, Error> {
