+++ crates/api/src/parser/terms/mapper/finishable.rs
@@ -1,4 +1,4 @@
-use crate::parser::Term;
+use crate::Term;
 
 pub(super) trait Finishable {
     fn finish(self) -> Vec<Term>;
