+++ crates/api/src/term.rs
