+++ crates/api/src/term/annotation_composable.rs
@@ -1,5 +1,5 @@
-use crate::parser::{
-    annotation_composition::{AnnotationComposable, AnnotationComposition},
+use crate::{
+    parser::annotation_composition::{AnnotationComposable, AnnotationComposition},
     Term,
 };
