+++ crates/api/src/term/partial_eq.rs
