+++ crates/api/src/unit/deref.rs
@@ -1,10 +1,10 @@
-use crate::{parser::Term, unit::Unit};
+use crate::Term;
 
 //-----------------------------------------------------------------------------
 // impl Deref
 //-----------------------------------------------------------------------------
 // TODO: Get rid of this in 1.0.
-impl ::std::ops::Deref for Unit {
+impl ::std::ops::Deref for crate::Unit {
     type Target = [Term];
 
     fn deref(&self) -> &[Term] {
